refactor: extract android discovery helpers
