+++ src/managers/android/mod.rs
@@ -279,6 +279,7 @@
 
 mod create;
 mod details;
+mod discovery;
 mod install;
 mod parser;
 mod sdk;
@@ -287,17 +288,12 @@ mod version;
 use crate::{
     constants::{
         commands, defaults,
-        env_vars::{self, HOME},
+        env_vars::HOME,
         files,
-        limits::{
-            ANDROID_COMMAND_PARTS_MINIMUM, STORAGE_MB_TO_GB_DIVISOR, SYSTEM_IMAGE_PARTS_REQUIRED,
-        },
+        limits::STORAGE_MB_TO_GB_DIVISOR,
         timeouts::{DEVICE_START_WAIT_TIME, DEVICE_STATUS_CHECK_DELAY},
     },
     managers::common::{DeviceConfig, DeviceManager},
-    models::device_info::{
-        ApiLevelInfo, DeviceCategory, DeviceInfo, DynamicDeviceConfig, DynamicDeviceProvider,
-    },
     models::{AndroidDevice, DeviceStatus},
     utils::command::CommandRunner,
     utils::command_executor::CommandExecutor,
@@ -305,7 +301,6 @@ use crate::{
 use anyhow::{Context, Result};
 use lazy_static::lazy_static;
 use regex::Regex;
-use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use tokio::fs;
@@ -454,670 +449,6 @@ impl AndroidManager {
             }
         }
         results
-    }
-
-    /// ```text
-    /// // Returns: {"Pixel_7_API_34" => "emulator-5554", "Pixel 7 API 34" => "emulator-5554"}
-    /// ```
-    pub async fn get_running_avd_names(&self) -> Result<HashMap<String, String>> {
-        let mut avd_map = HashMap::new();
-        let mut normalized_map = HashMap::new();
-
-        // Get list of running emulators
-        let adb_output = self
-            .command_executor
-            .run(
-                std::path::Path::new(commands::ADB),
-                &[commands::adb::DEVICES],
-            )
-            .await
-            .unwrap_or_default();
-
-        for line in adb_output.lines() {
-            if line.contains("emulator-") && line.contains("device") {
-                if let Some(emulator_id) = line.split_whitespace().next() {
-                    // log::debug!("Found emulator: {}", emulator_id);
-
-                    // Try multiple methods to get AVD name
-
-                    // Method 1: Try to get AVD name from boot properties (most reliable)
-                    if let Ok(boot_prop_output) = self
-                        .command_executor
-                        .run(
-                            Path::new(commands::ADB),
-                            &[
-                                "-s",
-                                emulator_id,
-                                "shell",
-                                "getprop",
-                                "ro.boot.qemu.avd_name",
-                            ],
-                        )
-                        .await
-                    {
-                        let avd_name = boot_prop_output.trim().to_string();
-                        // log::debug!("Method 1 - AVD name from boot property for {}: '{}'", emulator_id, avd_name);
-
-                        if !avd_name.is_empty() {
-                            avd_map.insert(avd_name.clone(), emulator_id.to_string());
-                            // Also store normalized version
-                            let normalized = avd_name.replace(' ', "_");
-                            if normalized != avd_name {
-                                normalized_map.insert(normalized, emulator_id.to_string());
-                            }
-                            continue;
-                        }
-                    }
-
-                    // Method 2: Try adb emu avd name (parse first line only)
-                    if let Ok(avd_name_output) = self
-                        .command_executor
-                        .run(
-                            std::path::Path::new(commands::ADB),
-                            &["-s", emulator_id, "emu", "avd", "name"],
-                        )
-                        .await
-                    {
-                        // Take only the first line to avoid "OK" or other status messages
-                        let avd_name = avd_name_output
-                            .lines()
-                            .next()
-                            .unwrap_or("")
-                            .trim()
-                            .to_string();
-
-                        // log::debug!("Method 2 - AVD name output for {}: '{}'", emulator_id, avd_name);
-
-                        // Check if this is a valid AVD name (not an error message)
-                        if !avd_name.is_empty()
-                            && !avd_name.contains("error")
-                            && !avd_name.contains("KO")
-                            && !avd_name.contains("unknown command")
-                            && avd_name != "OK"
-                        {
-                            avd_map.insert(avd_name.clone(), emulator_id.to_string());
-                            // Also store normalized version
-                            let normalized = avd_name.replace(' ', "_");
-                            if normalized != avd_name {
-                                normalized_map.insert(normalized, emulator_id.to_string());
-                            }
-                            continue;
-                        }
-                    }
-
-                    // Method 3: Try to get AVD name from kernel properties
-                    if let Ok(prop_output) = self
-                        .command_executor
-                        .run(
-                            Path::new(commands::ADB),
-                            &[
-                                "-s",
-                                emulator_id,
-                                "shell",
-                                "getprop",
-                                "ro.kernel.qemu.avd_name",
-                            ],
-                        )
-                        .await
-                    {
-                        let avd_name = prop_output.trim().to_string();
-                        // log::debug!("Method 3 - AVD name from kernel property for {}: '{}'", emulator_id, avd_name);
-
-                        if !avd_name.is_empty() {
-                            avd_map.insert(avd_name.clone(), emulator_id.to_string());
-                            // Also store normalized version
-                            let normalized = avd_name.replace(' ', "_");
-                            if normalized != avd_name {
-                                normalized_map.insert(normalized, emulator_id.to_string());
-                            }
-                            continue;
-                        }
-                    }
-
-                    // log::warn!("Could not determine AVD name for emulator {}", emulator_id);
-                }
-            }
-        }
-
-        // Merge normalized map into main map for any missing entries
-        for (normalized_name, serial) in normalized_map {
-            avd_map.entry(normalized_name).or_insert(serial);
-        }
-
-        Ok(avd_map)
-    }
-
-    /// List available Android targets (API levels) based on installed system images
-    pub async fn list_available_targets(&self) -> Result<Vec<(String, String)>> {
-        log::debug!("list_available_targets called");
-        // Try to load from cache first for faster response
-        if let Some(cached_targets) = crate::utils::ApiLevelCache::load_from_disk() {
-            let targets: Vec<(String, String)> = cached_targets
-                .api_levels
-                .into_iter()
-                .map(|api| {
-                    (
-                        api.api.to_string(),
-                        format!("API {} - {}", api.api, api.version),
-                    )
-                })
-                .collect();
-
-            if !targets.is_empty() {
-                log::debug!("Loaded {} API levels from cache", targets.len());
-                return Ok(targets);
-            }
-        }
-
-        // Cache miss - get actually installed system images
-        let installed_images = self.list_available_system_images().await?;
-
-        // Also track available tags/ABIs for each API level
-        let mut api_level_info: std::collections::HashMap<
-            String,
-            std::collections::HashSet<String>,
-        > = std::collections::HashMap::new();
-
-        let mut targets = std::collections::HashMap::new();
-
-        // Extract API levels and their available tags/ABIs from installed system images
-        for image in installed_images {
-            // Parse "system-images;android-XX;tag;abi" format
-            let parts: Vec<&str> = image.split(';').collect();
-            if parts.len() >= SYSTEM_IMAGE_PARTS_REQUIRED {
-                if let Some(api_level) = parts.get(1).and_then(|p| p.strip_prefix("android-")) {
-                    // Track available tags for this API level
-                    let tag_abi = format!("{};{}", parts[2], parts[3]);
-                    api_level_info
-                        .entry(api_level.to_string())
-                        .or_default()
-                        .insert(tag_abi);
-
-                    // Use simple version mapping to avoid recursion
-                    let api_num: u32 = api_level.parse().unwrap_or(0);
-                    let android_version = self.get_android_version_name(api_num);
-
-                    let display = format!("API {api_level} - {android_version}");
-                    targets.insert(api_level.to_string(), display);
-                }
-            }
-        }
-
-        // Convert to sorted vector
-        let mut result: Vec<(String, String)> = targets.into_iter().collect();
-
-        // Sort by API level (descending)
-        result.sort_by(|a, b| {
-            let api_a: u32 = a.0.parse().unwrap_or(0);
-            let api_b: u32 = b.0.parse().unwrap_or(0);
-            api_b.cmp(&api_a)
-        });
-
-        // Save to cache for future fast loading
-        let api_levels: Vec<crate::models::ApiLevel> = result
-            .iter()
-            .map(|(level_str, display)| {
-                let api: u32 = level_str.parse().unwrap_or(0);
-                let version = if let Some(dash_pos) = display.find(" - ") {
-                    display[dash_pos + 3..].to_string()
-                } else {
-                    format!("API {api}")
-                };
-                crate::models::ApiLevel {
-                    api,
-                    version,
-                    display_name: display.clone(),
-                    system_image_id: format!("android-{api}"),
-                    is_installed: true, // These are installed images
-                    variants: vec![],   // Empty for simple cache
-                }
-            })
-            .collect();
-
-        let cache = crate::utils::ApiLevelCache {
-            api_levels,
-            timestamp: std::time::SystemTime::now(),
-        };
-
-        if let Err(e) = cache.save_to_disk() {
-            log::warn!("Failed to save API level cache: {e}");
-        } else {
-            log::debug!("Saved {} API levels to cache", result.len());
-        }
-        log::debug!(
-            "list_available_targets completed, returning {} targets",
-            result.len()
-        );
-
-        Ok(result)
-    }
-
-    /// Lists all available Android device definitions from the SDK.
-    ///
-    /// Parses output from `avdmanager list device` to discover all device types
-    /// that can be used to create AVDs. Returns devices sorted by priority
-    /// (phones first, then tablets, etc.) and version.
-    ///
-    /// # Returns
-    /// Vector of tuples containing:
-    /// - Device ID (e.g., "pixel_7")
-    /// - Display name (e.g., "Pixel 7")
-    ///
-    /// # Device Priority
-    /// Devices are sorted using dynamic priority calculation based on:
-    /// 1. Category (phone > tablet > tv > wear > automotive)
-    /// 2. Version number (newer versions first)
-    /// 3. Manufacturer (Google/Pixel prioritized)
-    pub async fn list_available_devices(&self) -> Result<Vec<(String, String)>> {
-        let output = self
-            .command_executor
-            .run(&self.avdmanager_path, &["list", "device"])
-            .await
-            .context("Failed to list Android devices")?;
-
-        let mut devices = Vec::new();
-        let mut current_id = String::new();
-        let mut current_name = String::new();
-        let mut current_oem = String::new();
-
-        for line in output.lines() {
-            if let Some(caps) = ID_REGEX.captures(line) {
-                current_id = caps[1].to_string();
-            } else if let Some(caps) = NAME_REGEX.captures(line) {
-                current_name = caps[1].to_string();
-            } else if let Some(caps) = OEM_REGEX.captures(line) {
-                current_oem = caps[1].to_string();
-            } else if line.contains("-----") && !current_id.is_empty() {
-                // Create display string
-                let display = if !current_oem.is_empty() && current_oem != "Generic" {
-                    format!("{current_name} ({current_oem})")
-                } else {
-                    current_name.clone()
-                };
-
-                devices.push((current_id.clone(), display));
-
-                // Reset for next device
-                current_id.clear();
-                current_name.clear();
-                current_oem.clear();
-            }
-        }
-
-        // Don't forget the last device if exists
-        if !current_id.is_empty() {
-            let display = if !current_oem.is_empty() && current_oem != "Generic" {
-                format!("{current_name} ({current_oem})")
-            } else {
-                current_name.clone()
-            };
-            devices.push((current_id, display));
-        }
-
-        // If no devices found, log warning but don't add hardcoded fallbacks
-        if devices.is_empty() {
-            log::warn!(
-                "No Android device definitions found. Please check your Android SDK installation."
-            );
-        }
-
-        // Sort devices by priority (Pixel devices first, then by version)
-        devices.sort_by(|a, b| {
-            let priority_a = DynamicDeviceConfig::calculate_android_device_priority(&a.0, &a.1);
-            let priority_b = DynamicDeviceConfig::calculate_android_device_priority(&b.0, &b.1);
-            priority_a.cmp(&priority_b)
-        });
-
-        Ok(devices)
-    }
-
-    /// Dynamically determine device category
-    pub fn get_device_category(&self, device_id: &str, device_display: &str) -> String {
-        let combined = format!(
-            "{} {}",
-            device_id.to_lowercase(),
-            device_display.to_lowercase()
-        );
-
-        // Phone - most common device type
-        if combined.contains("phone") || 
-           combined.contains("pixel") && !combined.contains("fold") && !combined.contains("tablet") ||
-           combined.contains("galaxy") && !combined.contains("fold") && !combined.contains("tablet") ||
-           combined.contains("oneplus") ||
-           combined.contains("iphone") ||
-           // Determine by screen size (smartphone range: 5-6 inches)
-           Self::is_phone_size(&combined) ||
-           (combined.contains("pro") && !combined.contains("tablet") && !combined.contains("fold"))
-        {
-            return "phone".to_string();
-        }
-
-        // Tablet - tablet devices
-        if combined.contains("tablet")
-            || combined.contains("pad")
-            // Tablet size range: 10-13 inches
-            || Self::is_tablet_size(&combined)
-        {
-            return "tablet".to_string();
-        }
-
-        // Wear - wearable devices
-        if combined.contains("wear")
-            || combined.contains("watch")
-            || combined.contains("round") && !combined.contains("tablet")
-            || combined.contains("square") && !combined.contains("tablet")
-        {
-            return "wear".to_string();
-        }
-
-        // TV - television devices
-        if combined.contains("tv")
-            || combined.contains("1080p")
-            || combined.contains("4k")
-            || combined.contains("720p")
-        {
-            return "tv".to_string();
-        }
-
-        // Automotive - automotive devices
-        if combined.contains("auto") || combined.contains("car") || combined.contains("automotive")
-        {
-            return "automotive".to_string();
-        }
-
-        // Desktop - desktop/large screen devices
-        if combined.contains("desktop")
-            || combined.contains("foldable") && combined.contains("large")
-            // Large desktop size range: 15+ inches
-            || Self::is_desktop_size(&combined)
-        {
-            return "desktop".to_string();
-        }
-
-        // Default is phone (most common)
-        "phone".to_string()
-    }
-
-    /// Check if the device description indicates phone-sized screen (5-6 inches)
-    fn is_phone_size(combined: &str) -> bool {
-        if !combined.contains("inch") {
-            return false;
-        }
-
-        // Check for common phone sizes
-        for size in ["5", "6"] {
-            if combined.contains(size) {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Check if the device description indicates tablet-sized screen (10-13 inches)
-    fn is_tablet_size(combined: &str) -> bool {
-        if !combined.contains("inch") {
-            return false;
-        }
-
-        // Check for common tablet sizes
-        for size in ["10", "11", "12", "13"] {
-            if combined.contains(size) {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Check if the device description indicates desktop-sized screen (15+ inches)
-    fn is_desktop_size(combined: &str) -> bool {
-        if !combined.contains("inch") {
-            return false;
-        }
-
-        // Check for large desktop sizes
-        for size in ["15", "17"] {
-            if combined.contains(size) {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Get device list filtered by category
-    pub async fn list_devices_by_category(
-        &self,
-        category: Option<&str>,
-    ) -> Result<Vec<(String, String)>> {
-        let all_devices = self.list_available_devices().await?;
-
-        if let Some(filter_category) = category {
-            if filter_category == "all" {
-                return Ok(all_devices);
-            }
-
-            let filtered_devices: Vec<(String, String)> = all_devices
-                .into_iter()
-                .filter(|(id, display)| {
-                    let device_category = self.get_device_category(id, display);
-                    device_category == filter_category
-                })
-                .collect();
-
-            Ok(filtered_devices)
-        } else {
-            Ok(all_devices)
-        }
-    }
-
-    /// Dynamically get available skins for device from SDK
-    async fn get_available_skins_from_sdk(&self, _device_id: &str) -> Result<Vec<String>> {
-        let mut skins = Vec::new();
-
-        // Dynamically scan Android SDK skin directories
-        if let Ok(android_home) = std::env::var(env_vars::ANDROID_HOME) {
-            let android_path = std::path::PathBuf::from(&android_home);
-
-            // 1. Standard skin directory
-            let standard_skins = android_path.join("skins");
-            if standard_skins.exists() {
-                self.scan_skin_directory(&standard_skins, &mut skins).await;
-            }
-
-            // 2. Dynamically scan platform-specific skin directories
-            let platforms_dir = android_path.join("platforms");
-            if platforms_dir.exists() {
-                if let Ok(mut platform_entries) = fs::read_dir(&platforms_dir).await {
-                    while let Some(platform_entry) =
-                        platform_entries.next_entry().await.ok().flatten()
-                    {
-                        if let Ok(file_type) = platform_entry.file_type().await {
-                            if file_type.is_dir() {
-                                let platform_skins = platform_entry.path().join("skins");
-                                if platform_skins.exists() {
-                                    self.scan_skin_directory(&platform_skins, &mut skins).await;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // 3. Dynamically scan system image-specific skin directories
-            let system_images_dir = android_path.join("system-images");
-            if system_images_dir.exists() {
-                self.scan_system_images_for_skins(&system_images_dir, &mut skins)
-                    .await;
-            }
-        }
-
-        // 4. Get available device IDs from avdmanager (these are also skin candidates)
-        if let Ok(available_devices) = self.list_available_devices().await {
-            for (id, _display) in available_devices {
-                skins.push(id);
-            }
-        }
-
-        // Remove duplicates and sort
-        skins.sort();
-        skins.dedup();
-
-        Ok(skins)
-    }
-
-    /// Scan skin directory and collect skin names
-    async fn scan_skin_directory(&self, skin_dir: &std::path::Path, skins: &mut Vec<String>) {
-        if let Ok(mut entries) = fs::read_dir(skin_dir).await {
-            while let Some(entry) = entries.next_entry().await.ok().flatten() {
-                if let Ok(file_type) = entry.file_type().await {
-                    if file_type.is_dir() {
-                        if let Some(skin_name) = entry.file_name().to_str() {
-                            skins.push(skin_name.to_string());
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Recursively scan system image directories to find skins
-    async fn scan_system_images_for_skins(
-        &self,
-        system_images_dir: &std::path::Path,
-        skins: &mut Vec<String>,
-    ) {
-        if let Ok(mut api_entries) = fs::read_dir(system_images_dir).await {
-            while let Some(api_entry) = api_entries.next_entry().await.ok().flatten() {
-                if let Ok(file_type) = api_entry.file_type().await {
-                    if file_type.is_dir() {
-                        let api_dir = api_entry.path();
-                        if let Ok(mut tag_entries) = fs::read_dir(&api_dir).await {
-                            while let Some(tag_entry) =
-                                tag_entries.next_entry().await.ok().flatten()
-                            {
-                                if let Ok(file_type) = tag_entry.file_type().await {
-                                    if file_type.is_dir() {
-                                        let tag_dir = tag_entry.path();
-                                        if let Ok(mut abi_entries) = fs::read_dir(&tag_dir).await {
-                                            while let Some(abi_entry) =
-                                                abi_entries.next_entry().await.ok().flatten()
-                                            {
-                                                if let Ok(file_type) = abi_entry.file_type().await {
-                                                    if file_type.is_dir() {
-                                                        let skins_dir =
-                                                            abi_entry.path().join("skins");
-                                                        if skins_dir.exists() {
-                                                            self.scan_skin_directory(
-                                                                &skins_dir, skins,
-                                                            )
-                                                            .await;
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-impl DynamicDeviceProvider for AndroidManager {
-    async fn get_available_devices(&self) -> Result<Vec<DeviceInfo>> {
-        let devices = self.list_available_devices().await?;
-
-        let mut device_infos = Vec::new();
-        for (id, display_name) in devices {
-            // Use basic categorization to avoid deep async nesting
-            let category = DeviceCategory::Unknown; // Will be determined later if needed
-
-            // Extract OEM from display name for now (to avoid async issues)
-            let oem = if display_name.contains('(') && display_name.contains(')') {
-                let start = display_name.find('(').unwrap() + 1;
-                let end = display_name.find(')').unwrap();
-                Some(display_name[start..end].to_string())
-            } else {
-                None
-            };
-
-            device_infos.push(DeviceInfo {
-                id,
-                display_name,
-                oem,
-                category,
-            });
-        }
-
-        Ok(device_infos)
-    }
-
-    async fn get_available_api_levels(&self) -> Result<Vec<ApiLevelInfo>> {
-        let targets = self.list_available_targets().await?;
-
-        let mut api_infos = Vec::new();
-        for (api_level_str, display) in targets {
-            if let Ok(level) = api_level_str.parse::<u32>() {
-                // Extract version name from display (e.g., "API 34 - Android 14" -> "Android 14")
-                let version_name = if let Some(dash_pos) = display.find(" - ") {
-                    display[dash_pos + 3..].to_string()
-                } else {
-                    self.get_dynamic_android_version_name(level)
-                        .await
-                        .unwrap_or_else(|| format!("API {level}"))
-                };
-
-                // Get available tags for this API level
-                let available_tags = self
-                    .get_available_tags_for_api_level(level)
-                    .await
-                    .unwrap_or_default();
-
-                api_infos.push(ApiLevelInfo {
-                    level,
-                    version_name,
-                    available_tags,
-                });
-            }
-        }
-
-        // Sort by API level (newest first)
-        api_infos.sort_by(|a, b| b.level.cmp(&a.level));
-
-        Ok(api_infos)
-    }
-
-    async fn get_available_skins(&self, device_id: &str) -> Result<Vec<String>> {
-        // Dynamically get skins
-        self.get_available_skins_from_sdk(device_id).await
-    }
-
-    async fn get_device_priority(&self, device_id: &str) -> Result<u32> {
-        // Use basic priority calculation to avoid async issues
-        Ok(DynamicDeviceConfig::calculate_android_device_priority(
-            device_id, "",
-        ))
-    }
-}
-
-impl AndroidManager {
-    async fn get_available_tags_for_api_level(&self, api_level: u32) -> Result<Vec<String>> {
-        let images = self.list_available_system_images().await?;
-        let mut tags = HashSet::new();
-
-        for image in images {
-            if image.contains(&format!("android-{api_level}")) {
-                let parts: Vec<&str> = image.split(';').collect();
-                if parts.len() >= ANDROID_COMMAND_PARTS_MINIMUM {
-                    tags.insert(parts[2].to_string());
-                }
-            }
-        }
-
-        Ok(tags.into_iter().collect())
     }
 }
 
@@ -1483,6 +814,7 @@ impl crate::managers::common::UnifiedDeviceManager for AndroidManager {
 mod tests {
     use super::*;
     use crate::managers::common::DeviceConfig;
+    use crate::models::device_info::DynamicDeviceProvider;
     use crate::utils::command_executor::mock::MockCommandExecutor;
     use std::collections::HashMap;
     use std::env;
